+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Breaking changes
 
-* Change `Inspect` protocol for `A.Vector`
+* Change `Inspect` protocol for `A.Vector` and `A.OrdMap`
 
 ## v0.5.0 (2021-03-25)